refactored mobile top bar navigation for strike guides into a reusable component for other guides.
