+++ BAD-guild-website/src/data/strikeguidenavlinks.jsx
@@ -1,0 +1,19 @@
+//page navigation link ids and labels for strike guide
+
+const navLinks = [
+  { id: "Shiverpeaks Pass", label: "Shiverpeaks Pass" },
+  { id: "Fraenir of Jormag", label: "Fraenir of Jormag" },
+  { id: "Voice and Claw", label: "Voice and Claw" },
+  { id: "Whisper of Jormag", label: "Whisper of Jormag" },
+  { id: "Boneskinner", label: "Boneskinner" },
+  { id: "Cold War", label: "Cold War" },
+  { id: "Aetherblade Hideout", label: "Aetherblade Hideout" },
+  { id: "Xunlai Jade Junkyard", label: "Xunlai Jade Junkyard" },
+  { id: "Kaineng Overlook", label: "Kaineng Overlook" },
+  { id: "Harvest Temple", label: "Harvest Temple" },
+  { id: "Old Lion's Court", label: "Old Lion's Court" },
+  { id: "Cosmic Observatory", label: "Cosmic Observatory" },
+  { id: "Temple of Febe", label: "Temple of Febe" },
+];
+
+export default navLinks;
